fix(portals): break portal proxy/service type cycle; regenerate Swift protocol models
